Material: Move alphahash to separate shader chunks.
