+++ src/renderers/shaders/ShaderChunk/alphahash_fragment.glsl.js
@@ -1,0 +1,7 @@
+export default /* glsl */`
+#ifdef USE_ALPHAHASH
+
+	if ( diffuseColor.a < getAlphaHashThreshold( vPosition, alphaHashScale ) ) discard;
+
+#endif
+`;
